[Improvement] Separate stat for nerds helper
